build: create the program
I used Cargo to generate the starter program skeleton and the main.rs
program. I customized Cargo.toml with the program metadata.

I updated README.md with instructions on how to get started working with
the source code in a development container. I created a document
describing the software requirements for development using a development
container and for working with the source code locally outside of a
development container.